+++ engine/src/flutter/testing/dart/BUILD.gn
@@ -20,6 +20,7 @@ tests = [
   "gesture_settings_test.dart",
   "gpu_test.dart",
   "gradient_test.dart",
+  "high_bitrate_texture_test.dart",
   "http_allow_http_connections_test.dart",
   "http_disallow_http_connections_test.dart",
   "image_descriptor_test.dart",
